Refactor inference.rs into focused modules
Split the 3369-line inference.rs file into a clean module structure:

- inference/types.rs: TypeVarId, InferType, TypeVarAllocator (~175 lines)
- inference/constraint.rs: Constraint, Substitution (~230 lines)
- inference/unify.rs: UnifyResult, UnificationError, Unifier (~570 lines)
- inference/generate.rs: ConstraintContext, ConstraintGenerator (~1050 lines)
- inference/mod.rs: Module documentation and re-exports (~45 lines)

Each module is now well under 1100 lines and focuses on a single cohesive
responsibility. The module boundaries align with the logical phases of
Hindley-Milner type inference: types → constraints → unification → generation.

Public API unchanged - same re-exports from lib.rs.
